Split authentication policies into public definition, private implementation (#3536)
