+++ src/enclave/rpc_context.h
@@ -3,12 +3,12 @@
 #pragma once
 
 #include "ccf/claims_digest.h"
+#include "ccf/service/signed_req.h"
 #include "ccf/tx_id.h"
 #include "http/http_builder.h"
 #include "http/http_consts.h"
 #include "node/entities.h"
 #include "node/rpc/error.h"
-#include "service/tables/client_signatures.h"
 #include "service/tables/node_info_network.h"
 
 #include <llhttp/llhttp.h>
