+++ src/enclave/rpc_handler.h
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #pragma once
+#include "ccf/tx.h"
 #include "ds/buffer.h"
 #include "forwarder_types.h"
