+++ src/node/rpc/test/frontend_test_infra.h
@@ -3,6 +3,7 @@
 #define DOCTEST_CONFIG_IMPLEMENT
 #define DOCTEST_CONFIG_NO_EXCEPTIONS_BUT_WITH_ALL_ASSERTS
 #include "ccf/app_interface.h"
+#include "ccf/service/signed_req.h"
 #include "ccf/user_frontend.h"
 #include "crypto/rsa_key_pair.h"
 #include "ds/files.h"
@@ -14,7 +15,6 @@
 #include "node/rpc/serdes.h"
 #include "node_stub.h"
 #include "service/genesis_gen.h"
-#include "service/tables/client_signatures.h"
 
 #include <doctest/doctest.h>
 #include <iostream>
