+++ src/service/tables/governance_history.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "ccf/entity_id.h"
-#include "client_signatures.h"
+#include "ccf/service/signed_req.h"
 #include "service/map.h"
 
 namespace ccf
